remove pytorch dependency
