+++ python/ark/__init__.py
@@ -29,7 +29,6 @@ from .executor import Executor
 from .serialize import (
     save,
     load,
-    convert_state_dict,
 )
 
 from .model import (
@@ -85,7 +84,6 @@ __all__ = [
     "Executor",
     "save",
     "load",
-    "convert_state_dict",
     "Optimizer",
     "Trainer",
     "Model",
